requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
opencv-python-headless==4.9.0.80
google-generativeai==0.3.2
openai==1.12.0
python-multipart==0.0.9
jinja2==3.1.3
APScheduler==3.10.4
requests==2.31.0
PyYAML==6.0.1
SQLAlchemy==2.0.27
tenacity==8.2.3
scikit-image==0.22.0
numpy<2.0.0
git+https://github.com/FalkTannhaeuser/python-onvif-zeep.git
